// newgram-ui/src/app/services/models/destaque-create-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface DestaqueCreateDto {

  /**
   * Capa do destaque
   */
  capaDeDestaque?: Blob ;

  /**
   * Nome do destaque
   */
  nome?: string;

  /**
   * Lista de IDs das stories associadas ao destaque
   */
  storiesIds?: Array<number>;
}